Мой первый проект (Исправления №1).
