+++ src/main/java/Product.java
@@ -1,0 +1,9 @@
+public class Product {
+    String name = "без названия";
+    double cost = 0.00f;
+    Product() {}
+    Product(String name, double cost) {
+        this.name = name;
+        this.cost = cost;
+    }
+}
